mds: fix persistence of completed_requests
Previously we only persisted sessions with modified
completed_requests as a side effect of other modifications
to the sessionmap.  This was a bug, as it was quite
possible for completed_sessions to become very stale,
resulting in the possibility of incorrect double-apply
of client requests during replay.

To fix this, we add new state on LogSegment and
a new step in try_to_expire that will ensure that
any session that had requests completed during
the segment will be persisted in the sessionmap
before the segment is expired.

Signed-off-by: John Spray <john.spray@redhat.com>