ceph.spec.in: do not %ghost /var/run/ceph
Prior to this commit, we didn't install /var/run/ceph as a normal
directory. We used the %ghost directive and created the directory with
"mkdir" in %post.

It makes more sense for RPM to simply install this directory like the
rest of the %files. The purpose of this change is 1) simplicity and 2)
gain the ability for RPM to automatically set the ownership of the
directory to an alternate UID once Ceph itself gains unprivileged user
support.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>