WIP-5073 update 7 that can create and list buckets
This is far from over. Of course, "P3" are scattered around.
All XXX still need to be resolved. In particular, we don't have
a good way to access someone else's buckets, since we use slash
as a separator and that conflicts with URLs (unless we quote it).
But we're getting there.